Implemented generic static structes.
The structures now need to be defined in the tilset with a
resource-location and a unique identifier.
This unique identifier is now used as a type for the entity on creation
- and all unique entity-types are loaded into memory in the world.js
before this from the tileset(map).

So now new structures that are well defined (walls, ceiling, floor etc.)
Only need to be defined in the maps tileset with a unique ID and a
resouurce path - and they are automatically loaded and presented as
usable types when rendering the world!