Added precise hit, and implemented faction
Entities now all belong to a faction FRIENDLY or HOSTILE (default
HOSTILE).
Collisions between entities with same faction is ignored for now
(logged).

Results in:
Monsters can now be hit by e.g. a player weapon and be disposed - but
player can not attack himself. Next step is to implement health and
more advanced hit-triggers (take damage etc.. HIT its own category?).

This is the first iteration of this and I think it can be improved - but
for now it works! Height is not considered (but should be checked y pos
vs entity-height if it is actually a hit!)

Entities are now placed on the tiled-map and spawend in the correct
position like anything else! Need to also add the entities from the map
instead of having to be manually defined as legal entities for the
factory to be able to spawn them.

Entity factory returns a new entity based on the type and params given
(defined in constants and from the map (e.g. position)

Fixes #3 #4 #5