Implement HIT component type fixes #12
Also added a health action-component fixes #13
Also added a death action-component fixes #17

An entity can now have health and a HIT can now deal damage.
If the health of an entity reaches 0 or below a DEATH action is sendt
and an entity with the DEATH component has a death event!

Added sound, moved sound assets  and created all as instances in assets
to save some memory (instead of creating a new howler for each audio
component, it now references a componen in assets to e.g. play sound).

Also added music to the world! map-based music semi-integrated (need to
make this more automatic, but ill do that in #11)
This fixes #20

Stole some music from the awesome  gorescript-repo! Just for testing,
will add my own soon :)