Improved death-notice (only sends it once now)
Added pain and death animations as well as audio.
Now also finishes the death-animation before dispose is triggerd.
This is right now kind of hacky - i would like to use the onComplete on
the tween/animation to trigger a dispose on the parent - but cant get it
to trigger at the correct time (wait until actually complete) and be
able to pass parameters/functions to execute on complete from caller...