First iteration of HUD - displaying health
Is also adjusted when the players health changes.
- Need a better way for a component to 'assign' itself to the hud
Right now the components to be displated are sendt into the
hud-component manually in the parent.
- The update should also be a bit nicer - right now the receiver knows
the naming-convention, and based on the message type it chan target the
specific hud-component directly (should be more general!)

Works for now - have to see if it holds together when ammo and different
weapons are defined!!

closes #21