Improved HUD registration and updating
An event is now called on init for a component that has a HUD-ITEM that
registers it with its owner if it has a HUD-Component.

A generic UPDATE can now also be dispatched when the HUD-ITEMs value is
updated - and it will update the HUD. This makes it much cleaner to
register and update several/different hud items based on different
components for an entity.

Need to figure out a way so that the HUD-component does now have to be
defined first to get other components register-events

Also removed some deprecated parameter values from static entities (that
are no longer needed after the physics engine was removed)

Improves #21