World-fog now defined on the map closes #23
No longer hard-coded, can be changed on the loaded map and will update
in game!

Did not work as an entity as its a special component in the scene (not
loaded in the scene as a object3d), and has very simple parameters so
would be overkill to treat it and create a bunch of logic to load it as
e.g. an entity.

Works fine for now - nice to have it defined on the map, having to load
it in the game.js after the map is loaded is a bit of a pain but its
nicer than having to pass the entire scene into the world.js just to add
the fog... Can hopefully improve this!