Implemented inventory-component, and ammunition! closes #15
WeaponActionComponent is now more general and renamed to
InventoryComponent.
It now holds all weapons a playergas, any items and ammunition (and its
count).
Shooting a weapon now reduces available ammunition and the entity can no
longer shoot if it has no ammo!!

Also added ammunition factory and organized assets to behave with the
new structure.
The HUD is now also drawn properly with multiple items and according to
window height