Added triggers!
Now possible to define a door-trigger on a wall that makes it behave
like a door closes #47
- requires a link between wall and trigger (mark wall as triggerable),
and identical placement of trigger and triggerable wall. Also requires
collision on the position.

Now possible to add entity-spawners ! closes #53
- Triggers when a player walks over the trigger - spawns given quantity
of given entities on the location (configurable through the map, need a
different trigger for each spawn type)