Fixed animation bug closes #57
Also made the animation-component a bit cleaner, and easier to
understand!

Now behaves as expected

Also disposes tweens when animation-component parents are disposed. And
does not create a tween if no animation is being done (e.g if just
rotating, scale and position will be skipped) Should hopefully give a
slight performance boost